Update opcua-basics.js
Dear Sir or Madam,
I have checked the code and I have not seen support to work with ExtensionObject datatypes. 
I am interested in writing and reading structures from a PLC through OPCUA and nodered, however I could not manage to do so.

It would be grateful to add the support for this datatype and I would like to ask if is it possible. 
I currently can not add this contribution to this project since I have not the knowledge to do so.

Thanks in advance for your time, I am looking forward to hear your opinion about this proposal.

Yours Sincerely,

a.zabalo